feat: 추천, 비추천 기능 구현 완료

+++ src/main/java/Go/board/service/RecommendService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -29,9 +30,11 @@ public class RecommendService {
 
         Post post = postRepository.findById(articleId).orElse(null);
 
-        // 이미 좋아요 되어있으면 에러 반환
-        if(recommendRepository.findByMemberAndBoard(member, post).isPresent()) {
-            throw new Exception();
+        // 이미 좋아요 되어있으면 좋아요 취소
+        if(recommendRepository.findByMemberAndPost(member, post).isPresent()) {
+            Recommend recommend = recommendRepository.findByMemberAndPost(member,post).orElse(null);
+            recommendRepository.delete(recommend);
+            return member;
         }
 
         Recommend recommend = new Recommend();
@@ -45,17 +48,42 @@ public class RecommendService {
     }
 
     @Transactional
-    public Member notLike(int articleId, HttpServletRequest request) {
+    public Member dislike(int articleId, HttpServletRequest request) throws Exception {
         HttpSession session = request.getSession();
         Long memberId = (Long) session.getAttribute("memberId");
         Member member = memberRepository.findById(memberId).orElse(null);
 
         Post post = postRepository.findById(articleId).orElse(null);
 
-        Recommend recommend = recommendRepository.findByMemberAndBoard(member, post).orElse(null);
+        // 이미 싫어요 되어있으면 싫어요 취소
+        if(recommendRepository.findByMemberAndPost(member, post).isPresent()) {
+            Recommend recommend = recommendRepository.findByMemberAndPost(member,post).orElse(null);
+            recommendRepository.delete(recommend);
+            return member;
+        }
 
-        recommendRepository.delete(recommend);
+        Recommend recommend = new Recommend();
+        recommend.setMember(member);
+        recommend.setPost(post);
+        recommend.setRecommend(false);
+
+        recommendRepository.save(recommend);
 
         return member;
     }
+
+    public int getLike(int postId) {
+        // 좋아요 개수를 가져올 게시글 가져오기
+        Post post = postRepository.findById(postId).orElse(null);
+        List<Recommend> likePosts = recommendRepository.findByPostAndRecommend(post,true);
+        return likePosts.size();
+    }
+
+    public int getDislike(int postId) {
+        // 싫어요 개수를 가져올 게시글 가져오기
+        Post post = postRepository.findById(postId).orElse(null);
+        List<Recommend> likePosts = recommendRepository.findByPostAndRecommend(post,false);
+        return likePosts.size();
+    }
+
 }
